+++ config/bootstrap.php
@@ -34,8 +34,5 @@ if (!Configure::check('Assets.target')) {
 }
 
 if (!is_writeable(Configure::read('Assets.target'))) {
-    trigger_error(
-        sprintf('Directory %s not writeable', Configure::read('Assets.target')),
-        E_USER_ERROR
-    );
+    trigger_error(sprintf('Directory %s not writeable', Configure::read('Assets.target')), E_USER_ERROR);
 }