+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # 1.x branch
 ## 1.1 branch
+### 1.1.3
+* added `AssetNotFoundException` class. This exception is thrown when you
+    request an asset that is not available;
+* updated for CakePHP 3.4.
+
 ### 1.1.2
 * fixed bug. Urls no longer contain the type of asset, but only the filename.